docs: fixes conversion of ghfm admonitions
The GitHub Flavored Markdown Spec allows for admonitions that are
presently not able to be parsed by the Myst parser used by our
Sphinx and ReadTheDocs solution. This commit implements a stopgap
resolution for this issue, until Myst parser fixes this behavior and
ReadTheDocs migrates to that Myst parser version.

Adds a pre-commit hook to enforce styling of GitHub flavored markdown
admonitions to match the needs of the conversion function.

Changelog: fixed - Fixed conversion of GHFM admonitions to RST